requirements.txt
altair==4.1.0
APScheduler==3.6.3
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
cffi==1.14.2
click==7.1.2
constantly==15.1.0
cryptography==3.0
daphne==2.5.0
Django==3.1
django-lockdown==3.0.0
entrypoints==0.3
Flask==1.1.2
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.21
numpy==1.19.1
pandas==1.1.0
protobuf==3.13.0
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.5.3
pyOpenSSL==19.1.0
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
schedule==0.6.0
service-identity==18.1.0
six==1.15.0
SQLAlchemy==1.2.16
sqlparse==0.3.1
toml==0.10.1
toolz==0.10.0
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
tzlocal==2.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
zope.interface==5.1.0
